Clean up old code, make some moves/renames
- cleaned up imports
- removed .base, .testers, .mixins
- removed unused functions in .raml_utils, .utils
- renamed named_params_schema to named_params_to_json_schema
- moved .raml_utils functions to .raml and removed .raml_utils
- moved ra.api module to ra.dsl to not clash with ra.api() function
- renamed ra.dsl.API to ra.dsl.APISuite to be more descriptive
