+++ ra/raml.py
@@ -4,6 +4,7 @@ as well as wrapper classes for the main ramlfications types to make
 them more pleasant to work with.
 """
 import collections
+import re
 import six
 import ramlfications
 import wrapt
@@ -13,6 +14,83 @@ from .utils import list_to_dict
 def parse(raml_path_or_string):
     root = ramlfications.parse(raml_path_or_string)
     return RootNode(root)
+
+
+def is_raml(s):
+    return s.startswith("#%RAML")
+
+
+STRIP_DYNAMIC = re.compile(r'/\{.*?}')
+
+def resource_name_from_path(path, singularize=True):
+    """Returns a (possibly nested) resource name for an API path).
+
+    This function will try to treat collection and item paths the same,
+    singularizing the collection name if singularize=True (default).
+    Nested resources on the item are appended to make a dotted name
+    for the subresource.
+
+    For example, both "/users" and "/users/{username}" return 'user',
+    while '/users/{username}/profile' returns 'user.profile'.
+
+    Attribute resources like "/users/{username}/settings" will return
+    'user.settings' despite not being model object (it's treated the same
+    by Ra).
+    """
+    parts = STRIP_DYNAMIC.sub('', path.strip('/')).split('/')
+    if singularize:
+        import inflection
+        parts = (inflection.singularize(part) for part in parts)
+    return '.'.join(parts)
+
+
+def uri_args_from_example(resource_node):
+    """Recursively determine example values for any URI args
+    in the resource path.
+    """
+    uri_args = {}
+    if resource_node.parent:
+        uri_args = uri_args_from_example(resource_node.parent)
+
+    if resource_node.uri_params is None:
+        return uri_args
+
+    for name, param in six.iteritems(resource_node.uri_params):
+        uri_args[name] = param.example
+
+    return uri_args
+
+
+def resource_full_path(path, parent=None):
+    if parent is None:
+        return path
+    return parent.path + path
+
+
+def named_params_to_json_schema(params):
+    """ Convert RAML "named parameters" to JSON schema params.
+
+    Only params that participate in JSON schema validation
+    are translated.
+
+    Does not support:
+        type: file
+    """
+    schema = {'type': params['type']}
+
+    if schema['type'] == 'date':
+        schema['type'] = 'string'
+        if 'pattern' not in params:
+            schema['format'] = 'date-time'
+
+    optional = ('enum', 'minLength', 'maxLength', 'minimum',
+                'maximum', 'required', 'pattern', 'default')
+
+    for param in optional:
+        if param in params:
+            schema[param] = params[param]
+
+    return schema
 
 
 class RootNode(wrapt.ObjectProxy):
